Fix price serialization for MercadoLibre sync service (#57)
* Normalize price handling to ensure JSON-serializable float values in MercadoLibre sync service.

* Fix test assertion to use JSON-serializable float for price instead of `Decimal`.

+++ aiecommerce/services/mercadolibre_publisher_impl/sync_service.py
@@ -16,10 +16,11 @@ class MercadoLibreSyncService:
         self._price_engine = MercadoLibrePriceEngine()
         self._stock_engine = MercadoLibreStockEngine()
 
-    def _normalize_price(self, value: Decimal | float | int | None) -> Decimal | None:
+    def _normalize_price(self, value: Decimal | float | int | None) -> float | None:
         if value is None:
             return None
-        return Decimal(str(value))
+        # Ensure price values are JSON-serializable for API requests.
+        return float(Decimal(str(value)))
 
     def sync_listing(self, listing: MercadoLibreListing, dry_run: bool = False, force: bool = False) -> bool:
         """
